requirements.txt
asgiref
Django
easy-thumbnails==2.7.1
Pillow
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
six
django-allauth
django-modeltranslation